+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/domain/OnItemClickListener.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.code_check.domain
+
+import jp.co.yumemi.android.code_check.model.Item
+
+interface OnItemClickListener{
+    fun itemClick(item: Item)
+}
