+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/one/OneFragment.kt
@@ -3,21 +3,17 @@
  */
 package jp.co.yumemi.android.code_check.ui.one
 
-import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.view.inputmethod.InputMethodManager
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.LifecycleCoroutineScope
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.*
 import jp.co.yumemi.android.code_check.R
 import jp.co.yumemi.android.code_check.databinding.FragmentOneBinding
+import jp.co.yumemi.android.code_check.domain.OnItemClickListener
 import jp.co.yumemi.android.code_check.model.Item
+import jp.co.yumemi.android.code_check.ui.one.adapter.CustomAdapter
 
 class OneFragment: Fragment(R.layout.fragment_one){
 
@@ -33,7 +29,7 @@ class OneFragment: Fragment(R.layout.fragment_one){
         val dividerItemDecoration = DividerItemDecoration(context, layoutManager.orientation)
         val viewModel = OneViewModel()
         val adapter = CustomAdapter(
-            object : CustomAdapter.OnItemClickListener {
+            object : OnItemClickListener {
                 override fun itemClick(item: Item) {
                     viewModel.goToRepositoryFragment(item, findNavController())
                 }
