+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/one/adapter/ViewHolder.kt
@@ -1,0 +1,6 @@
+package jp.co.yumemi.android.code_check.ui.one.adapter
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class ViewHolder(view: View): RecyclerView.ViewHolder(view)
